requirements.txt
openai
iso639-lang
pydub
click